requirements.txt
fastapi
uvicorn
pydantic
transformers
pillow
requests
groq
python-dotenv
torch
torchvision
google-generativeai